Fix some more tests
