+++ security-jwt/src/main/java/io/micronaut/security/token/jwt/endpoints/KeysControllerConfigurationProperties.java
@@ -19,7 +19,6 @@ import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.core.util.StringUtils;
 import io.micronaut.security.config.SecurityConfigurationProperties;
-import io.micronaut.security.endpoints.AbstractControllerConfigurationProperties;
 
 /**
  * Configures the {@link io.micronaut.security.token.jwt.endpoints.KeysController}.
@@ -29,7 +28,7 @@ import io.micronaut.security.endpoints.AbstractControllerConfigurationProperties
  */
 @Requires(property = KeysControllerConfigurationProperties.PREFIX + ".enabled", notEquals = StringUtils.FALSE, defaultValue = StringUtils.TRUE)
 @ConfigurationProperties(KeysControllerConfigurationProperties.PREFIX)
-public class KeysControllerConfigurationProperties extends AbstractControllerConfigurationProperties implements KeysControllerConfiguration {
+public class KeysControllerConfigurationProperties implements KeysControllerConfiguration {
 
     public static final String PREFIX = SecurityConfigurationProperties.PREFIX + ".endpoints.keys";
 
@@ -45,8 +44,21 @@ public class KeysControllerConfigurationProperties extends AbstractControllerCon
     @SuppressWarnings("WeakerAccess")
     public static final String DEFAULT_PATH = "/keys";
 
-    protected boolean enabled = DEFAULT_ENABLED;
-    protected String path = DEFAULT_PATH;
+    private boolean enabled = DEFAULT_ENABLED;
+    private String path = DEFAULT_PATH;
+
+    /**
+     * @return true if you want to enable the {@link io.micronaut.security.token.jwt.endpoints.KeysController}.
+     */
+    @Override
+    public boolean isEnabled() {
+        return this.enabled;
+    }
+
+    @Override
+    public String getPath() {
+        return this.path;
+    }
 
     /**
      * Enables {@link io.micronaut.security.token.jwt.endpoints.KeysController}. Default value {@value #DEFAULT_ENABLED}.
