--- security-oauth2/src/main/java/io/micronaut/security/oauth2/endpoint/AbstractPersistableConfigurationProperties.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017-2022 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.security.oauth2.endpoint;
-
-import java.util.Optional;
-
-/**
- * Base implementation for {@link PersistableConfiguration} implementations.
- *
- * @author Álvaro Sánchez-Mariscal
- */
-public abstract class AbstractPersistableConfigurationProperties implements PersistableConfiguration {
-
-    protected String persistence;
-    protected boolean enabled;
-
-    @Override
-    public Optional<String> getPersistence() {
-        return Optional.ofNullable(persistence);
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return enabled;
-    }
-
-}
