+++ security-oauth2/src/main/java/io/micronaut/security/oauth2/endpoint/authorization/state/DefaultStateConfiguration.java
@@ -17,8 +17,9 @@ package io.micronaut.security.oauth2.endpoint.authorization.state;
 
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.security.oauth2.configuration.OauthConfigurationProperties;
-import io.micronaut.security.oauth2.endpoint.AbstractPersistableConfigurationProperties;
 import io.micronaut.security.oauth2.endpoint.authorization.state.validation.StateValidationConfiguration;
+
+import java.util.Optional;
 
 /**
  * Configuration properties implementation of state validation configuration.
@@ -27,7 +28,7 @@ import io.micronaut.security.oauth2.endpoint.authorization.state.validation.Stat
  * @since 1.2.0
  */
 @ConfigurationProperties(DefaultStateConfiguration.PREFIX)
-public class DefaultStateConfiguration extends AbstractPersistableConfigurationProperties implements StateValidationConfiguration {
+public class DefaultStateConfiguration implements StateValidationConfiguration {
 
     public static final String PREFIX = OauthConfigurationProperties.PREFIX + ".state";
 
@@ -38,8 +39,13 @@ public class DefaultStateConfiguration extends AbstractPersistableConfigurationP
     public static final boolean DEFAULT_ENABLED = true;
     private static final String DEFAULT_PERSISTENCE = "cookie";
 
-    protected String persistence = DEFAULT_PERSISTENCE;
-    protected boolean enabled = DEFAULT_ENABLED;
+    private String persistence = DEFAULT_PERSISTENCE;
+    private boolean enabled = DEFAULT_ENABLED;
+
+    @Override
+    public Optional<String> getPersistence() {
+        return Optional.ofNullable(persistence);
+    }
 
     /**
      * Sets the mechanism to persist the state for later retrieval for validation.
@@ -49,6 +55,11 @@ public class DefaultStateConfiguration extends AbstractPersistableConfigurationP
      */
     public void setPersistence(String persistence) {
         this.persistence = persistence;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return enabled;
     }
 
     /**
